chore: move `prism` config into `bin` dir
